requirements.txt
aiogram==3.13.1
SQLAlchemy==2.0.36
alembic==1.14.0
python-dotenv==1.0.1
aiosqlite==0.19.0
